Safari-611.1.10/LayoutTests/crypto/subtle/ecdsa-import-key-sign-p256-expected.txt

Test ECDSA signing operation with a P-256 private key

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS signature.byteLength is 64
PASS successfullyParsed is true

TEST COMPLETE
